Refactored NumberMask to support 64-bits numeric serialization;

+++ binary-proto/src/main/java/com/jd/blockchain/binaryproto/impl/AbstractDynamicValueConverter.java
@@ -39,7 +39,7 @@ public abstract class AbstractDynamicValueConverter implements DynamicValueConve
 	}
 	
 	protected int readSize(BytesInputStream bytesStream) {
-		return SIZE_HEAD.resolveMaskedNumber(bytesStream);
+		return (int)SIZE_HEAD.resolveMaskedNumber(bytesStream);
 	}
 
 }
